Get the address space map of a process

Returns the address space map of a process as a data frame.

procmap_get(..., as_tibble = NULL)

Arguments

...

Reserved for future extensions, must be empty.

as_tibble

When using in a package, set to TRUE to return a tibble::tibble. This requires the tibble package to be installed. The default returns a tibble if the package is installed, otherwise a data frame.

Value

A data frame or tibble, depending on the as_tibble argument.

Examples

procmap_get()
#> # A tibble: 548 × 6
#>    from             to               perms offset   inode  pathname             
#>    <chr>            <chr>            <chr> <chr>    <chr>  <chr>                
#>  1 0000556210c2c000 0000556210c2d000 r--p  00000000 807941 "/opt/R/4.5.1/lib/R/…
#>  2 0000556210c2d000 0000556210c2e000 r-xp  00001000 807941 "/opt/R/4.5.1/lib/R/…
#>  3 0000556210c2e000 0000556210c2f000 r--p  00002000 807941 "/opt/R/4.5.1/lib/R/…
#>  4 0000556210c2f000 0000556210c30000 r--p  00002000 807941 "/opt/R/4.5.1/lib/R/…
#>  5 0000556210c30000 0000556210c31000 rw-p  00003000 807941 "/opt/R/4.5.1/lib/R/…
#>  6 000055623f3e8000 000055624608e000 rw-p  00000000 0      "[heap]"             
#>  7 00007f096c000000 00007f096c021000 rw-p  00000000 0      ""                   
#>  8 00007f096c021000 00007f0970000000 ---p  00000000 0      ""                   
#>  9 00007f0973179000 00007f097317e000 r--p  00000000 78746  "/usr/lib/x86_64-lin…
#> 10 00007f097317e000 00007f09731f6000 r-xp  00005000 78746  "/usr/lib/x86_64-lin…
#> # ℹ 538 more rows