Get the address space map of a process

Returns the address space map of a process as a data frame.

procmap_get(..., as_tibble = NULL)

Arguments

...
Reserved for future extensions, must be empty.

as_tibble
When using in a package, set to TRUE to return a tibble::tibble. This requires the tibble package to be installed. The default returns a tibble if the package is installed, otherwise a data frame.

Value

A data frame or tibble, depending on the as_tibble argument.

Examples

procmap_get()
#> # A tibble: 499 × 6
#>    from           to             perms offset  inode  pathname                  
#>    <chr>          <chr>          <chr> <chr>   <chr>  <chr>                     
#>  1 000055dc893a4… 000055dc893a5… r-xp  000000… 41503… "/usr/lib/R/bin/exec/R"   
#>  2 000055dc895a4… 000055dc895a5… r--p  000000… 41503… "/usr/lib/R/bin/exec/R"   
#>  3 000055dc895a5… 000055dc895a6… rw-p  000010… 41503… "/usr/lib/R/bin/exec/R"   
#>  4 000055dc89e67… 000055dc900ac… rw-p  000000… 0      "[heap]"                  
#>  5 00007f87c0000… 00007f87c0021… rw-p  000000… 0      ""                        
#>  6 00007f87c0021… 00007f87c4000… ---p  000000… 0      ""                        
#>  7 00007f87c4fa5… 00007f87c4fba… r-xp  000000… 56974… "/usr/lib/R/modules/inter…
#>  8 00007f87c4fba… 00007f87c51b9… ---p  000150… 56974… "/usr/lib/R/modules/inter…
#>  9 00007f87c51b9… 00007f87c51ba… r--p  000140… 56974… "/usr/lib/R/modules/inter…
#> 10 00007f87c51ba… 00007f87c51bb… rw-p  000150… 56974… "/usr/lib/R/modules/inter…
#> # … with 489 more rows